indexer/requirements.txt
anyio==4.4.0
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jsonalias==0.1.1
jsonrpcclient==4.0.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sniffio==1.3.1
solana==0.34.3
solders==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==11.0.3
